+++ flows/definition/flow.go
@@ -9,6 +9,7 @@ import (
 	"github.com/nyaruka/goflow/envs"
 	"github.com/nyaruka/goflow/excellent/types"
 	"github.com/nyaruka/goflow/flows"
+	"github.com/nyaruka/goflow/flows/definition/legacy"
 	"github.com/nyaruka/goflow/flows/definition/migrations"
 	"github.com/nyaruka/goflow/flows/inspect"
 	"github.com/nyaruka/goflow/utils"
@@ -341,10 +342,37 @@ func IsSpecVersionSupported(ver *semver.Version) bool {
 	return ver.Major() >= 13 && ver.Major() <= CurrentSpecVersion.Major()
 }
 
-// ReadFlow reads a single flow definition from the passed in byte array
-func ReadFlow(data json.RawMessage) (flows.Flow, error) {
+// MigrationConfig configures how flow migrations are handled
+type MigrationConfig struct {
+	BaseMediaURL string
+}
+
+// ReadFlow a flow definition from the passed in byte array, migrating it to the spec version of the engine if necessary
+func ReadFlow(data json.RawMessage, migrationConfig *MigrationConfig) (flows.Flow, error) {
+	// try to read header (uuid, name, spec_version)
 	header := &flowHeader{}
-	if err := utils.UnmarshalAndValidate(data, header); err != nil {
+	err := utils.UnmarshalAndValidate(data, header)
+
+	if err != nil {
+		// could this be a legacy definition?
+		if legacy.IsPossibleDefinition(data) {
+			if migrationConfig == nil {
+				return nil, errors.New("unable to migrate what appears to be a legacy definition without a migration config")
+			}
+
+			// try to migrate it forwards to 13.0.0
+			var err error
+			data, err = legacy.MigrateDefinition(data, migrationConfig.BaseMediaURL)
+			if err != nil {
+				return nil, errors.Wrap(err, "error migrating what appears to be a legacy definition")
+			}
+		}
+
+		// try reading header again
+		err = utils.UnmarshalAndValidate(data, header)
+	}
+
+	if err != nil {
 		return nil, errors.Wrap(err, "unable to read flow header")
 	}